Fix broken breadcrumbs
Was including all kinds of links.
